updated projects layout
- created two components:
CurrentProjects and PastProjects
