+++ src/components/projects/Projects.css
@@ -40,7 +40,6 @@
 }
 
 .project-card {
-  margin-left: 135px;
   text-align: left;
   color: white;
   padding-top: 30px;
